backend/pyproject.toml
[project]
name = "invitation-backend"
version = "0.1.0"
description = "Backend for invitation website"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
    "sqlalchemy>=2.0.41,<3.0.0",
    "alembic>=1.16.0,<2.0.0",
    "asyncpg>=0.30.0,<1.0.0",
    "PyJWT>=2.10.0,<3.0.0",
    "pydantic-settings>=2.10.0,<3.0.0",
    "pwdlib[argon2]>=0.3.0,<1.0.0",
    "boto3>=1.39.0,<2.0.0",
    "python-multipart>=0.0.20,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.0,<1.0.0",
    "pytest>=8.4.0,<9.0.0",
    "pytest-asyncio>=1.1.0,<2.0.0",
]

[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
